% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfantasy_1.R
\docType{data}
\name{finalfantasy_1}
\alias{finalfantasy_1}
\title{Final Fantasy I Enemy Stats}
\format{
A data frame with 128 rows and
28 variables:
\subsection{Basic Information}{

\describe{
\item{name}{Character, name of enemy.}
\item{bestiary}{Integer, location in bestiary.}
\item{exp}{Integer, amount of experiance gained when defeated.}
\item{gil}{Integer, amount of gil gained when defeated.}
}
}

\subsection{Stats}{

\describe{
\item{hp}{Integer,maximum HP.}
\item{defense}{Integer, reduces the amount of physical damage received.}
\item{evasion}{Integer, affects the success rate of dodging enemy attacks.}
\item{magic_defense}{Integer, reduces the amount of magical damage
received.}
\item{attack}{Integer, affects the physical damage caused.}
\item{accuracy}{Integer, affects a character's chance to hit with physical
attacks.}
\item{hits}{Integer, affects the maximum number of hits player characters
can land.}
\item{crit}{Integer, determines the frequency with which they land a
critical hit.}
}
}

\subsection{Elemental affinities}{

\describe{
\item{fire}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{ice}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{lightning}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{earth}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{spirit}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{poison_element}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{time}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{instant_death}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{dia}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
}
}

\subsection{Status affinities}{

\describe{
\item{silence}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{sleep}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{paralysis}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{darkness}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{poison_status}{Character, its affinity, can be one of \code{"Misses"},
\code{"None"}, \code{"Resists"}, \code{"Weak"}.}
\item{stone}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
\item{confuse}{Character, its affinity, can be one of \code{"Misses"}, \code{"None"},
\code{"Resists"}, \code{"Weak"}.}
}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy)#Pixel_Remaster}

\url{https://finalfantasy.fandom.com/wiki/Final_Fantasy_stats}
}
\usage{
finalfantasy_1
}
\description{
A dataset of the stats of the enemies in Final Fantasy. Specifically those
appearning in the Pixel Remaster edition.
}
\seealso{
\link{finalfantasy_1_info}
}
\keyword{datasets}
