% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfantasy_1.R
\docType{data}
\name{finalfantasy_1_info}
\alias{finalfantasy_1_info}
\title{Final Fantasy I Enemy Additional Information}
\format{
A data frame with 128 rows and
7 variables:

\describe{
\item{name}{Character, name of enemy.}
\item{bestiary}{Integer, location in bestiary.}
\item{type}{Character, enemy type, unknown are lsited as \code{NA}.}
\item{location}{Character, comma separated list of location names.}
\item{drop}{Character, comma separated list of drops.}
\item{abilities}{Character, comma separated list of abilities.}
\item{url}{Character, url of the enemy.}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy)#Pixel_Remaster}
}
\usage{
finalfantasy_1_info
}
\description{
A dataset of additional information of the enemies in Final Fantasy.
Specifically those appearning in the Pixel Remaster edition.
}
\seealso{
\link{finalfantasy_1}
}
\keyword{datasets}
